requirements.txt
nibabel
monai
hydra-core
tensorboard
tqdm
omegaconf